requirements.txt
torch>=2.2.0
transformers>=4.31.0
accelerate>=0.28.0
peft>=0.4.0
einops>=0.7.0
sentencepiece>=0.2.0

numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.12.0
statsmodels>=0.14.0

fredapi>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0

matplotlib>=3.7.0
seaborn>=0.13.0

pyyaml>=6.0
tqdm>=4.65.0
